fixed failing test, removed old theme file and added transition to default theme
