+++ src/styles/theme.ts
@@ -1,105 +1,78 @@
-const palette = {
-  common: {
-    black: '#000',
-    white: '#fff',
-  },
-  primary: {
-    light: '#CEEAE7',
-    main: '#8fdecb',
-    dark: '#8FDDCA',
-    contrastText: '#4C4F5A',
-    400: '#EDF7F6',
-    600: '#68baa6',
-  },
-  secondary: {
-    light: '#FEFCDD',
-    main: '#FFEFD2',
-    dark: '#ba9147',
-    contrastText: '#4C4F5A',
-    400: '#fff7e8',
-  },
-  error: {
-    light: '#f47f98',
-    main: '#EF4836',
-    dark: '#d32f2f',
-    contrastText: '#fff',
-  },
-  text: {
-    primary: '#4C4F5A',
-    secondary: 'rgba(0, 0, 0, 0.54)',
-    disabled: 'rgba(0, 0, 0, 0.38)',
-    hint: 'rgba(0, 0, 0, 0.38)',
-  },
-  action: {
-    disabled: 'rgba(0, 0, 0, 0.07)',
-    hover: 'rgba(0, 0, 0, 0.04)',
-    disabledBackground: 'rgba(0, 0, 0, 0.05)',
-  },
-};
+import { createGlobalStyle } from 'styled-components';
 
-const spacing = {
-  padding: {
-    sm: 10,
-    md: 20,
-    lg: 30,
-  },
-};
+export const GlobalStyle = createGlobalStyle`
+  :root {
+    --palette-common-white: #ffffff;
+    --palette-common-black: #000000;
+    
+    --palette-primary-light: #CEEAE7;
+    --palette-primary-main: #8fdecb;
+    --palette-primary-dark: #8FDDCA;
+    --palette-primary-contrastText: #4C4F5A;
+    --palette-primary-400: #EDF7F6;
+    --palette-primary-600: #68baa6;
+    
+    --palette-secondary-light: #FEFCDD;
+    --palette-secondary-main: #FFEFD2;
+    --palette-secondary-dark: #ba9147;
+    --palette-secondary-contrastText: #4C4F5A;
+    --palette-secondary-400: #fff7e8;
+    
+    --palette-error-light: #f47f98;
+    --palette-error-main: #EF4836;
+    --palette-error-dark: #d32f2f;
+    --palette-error-contrastText: #ffffff;
+    
+    --palette-text-primary: #4C4F5A;
+    --palette-text-secondary: rgba(0, 0, 0, 0.54);
+    --palette-text-disabled: rgba(0, 0, 0, 0.38);
+    --palette-text-hint: rgba(0, 0, 0, 0.38);
+    
+    --palette-action-disabled: rgba(0, 0, 0, 0.07);
+    --palette-action-hover: rgba(0, 0, 0, 0.04);
+    --palette-action-disabledBackground: rgba(0, 0, 0, 0.05);
+    
+    --border-radius-small: 10px;
+    --border-radius-default: 40px;
 
-const borderRadius = {
-  small: '10px',
-  default: '40px',
-};
+    --html-font-size: 16;
+    --default-font-size: 14;
+    --font-weight-light: 300;
+    --font-weight-regular: 400;
+    --font-weight-medium: 500;
+    --font-weight-bold: 700;
+    
+    --shadow-none: none;
+    --shadow-01: 0px 2px 1px -1px rgba(0,0,0,0.2),0px 1px 1px 0px rgba(0,0,0,0.14),0px 1px 3px 0px rgba(0,0,0,0.12);
+    --shadow-02: 0px 3px 1px -2px rgba(0,0,0,0.2),0px 2px 2px 0px rgba(0,0,0,0.14),0px 1px 5px 0px rgba(0,0,0,0.12);
+    --shadow-03: 0px 3px 3px -2px rgba(0,0,0,0.2),0px 3px 4px 0px rgba(0,0,0,0.14),0px 1px 8px 0px rgba(0,0,0,0.12);
+    --shadow-04: 0px 2px 4px -1px rgba(0,0,0,0.2),0px 4px 5px 0px rgba(0,0,0,0.14),0px 1px 10px 0px rgba(0,0,0,0.12);
+    --shadow-05: 0px 3px 5px -1px rgba(0,0,0,0.2),0px 5px 8px 0px rgba(0,0,0,0.14),0px 1px 14px 0px rgba(0,0,0,0.12);
+    --shadow-06: 0px 3px 5px -1px rgba(0,0,0,0.2),0px 6px 10px 0px rgba(0,0,0,0.14),0px 1px 18px 0px rgba(0,0,0,0.12);
+    --shadow-07: 0px 4px 5px -2px rgba(0,0,0,0.2),0px 7px 10px 1px rgba(0,0,0,0.14),0px 2px 16px 1px rgba(0,0,0,0.12);
+    --shadow-08: 0px 5px 5px -3px rgba(0,0,0,0.2),0px 8px 10px 1px rgba(0,0,0,0.14),0px 3px 14px 2px rgba(0,0,0,0.12);
+    
+    --ease-out:cubic-bezier(0.0, 0, 0.2, 1); 
+    --ease-in:cubic-bezier(0.4, 0, 1, 1); 
+    
+    --spacing-sm: 10;
+    --spacing-md: 20;
+    --spacing-lg: 30;
+  }
+  
+  button {
+    font-family: 'Roboto', sans-serif;
+    font-weight: 500;
+    font-size: 0.875rem;
+    line-height: 1.75;
+    letter-spacing: 0.02857em;
+    text-transform: uppercase;
+  }
+`;
 
-/**
- * Typography
- */
-const defaultFontSize: number = 14;
-const htmlFontSize: number = 16;
-const coefficient: number = defaultFontSize / 14;
-const getRemSize = (size: number) =>
-  `${(size / htmlFontSize) * coefficient}rem`;
-
-const typography = {
-  htmlFontSize: htmlFontSize,
-  remSize: (fontSize: number) => getRemSize(fontSize),
-  defaultFontSize: defaultFontSize,
-  fontWeightLight: 300,
-  fontWeightRegular: 400,
-  fontWeightMedium: 500,
-  fontWeightBold: 700,
-  button: {
-    fontFamily: '"Roboto", sans-serif',
-    fontWeight: 500,
-    fontSize: '0.875rem',
-    lineHeight: 1.75,
-    letterSpacing: '0.02857em',
-    textTransform: 'uppercase',
-  },
-};
-
-const easing = {
-  easeOut: 'cubic-bezier(0.0, 0, 0.2, 1)',
-  easeIn: 'cubic-bezier(0.4, 0, 1, 1)',
-};
-
-const shadows = [
-  'none',
-  '0px 2px 1px -1px rgba(0,0,0,0.2),0px 1px 1px 0px rgba(0,0,0,0.14),0px 1px 3px 0px rgba(0,0,0,0.12)',
-  '0px 3px 1px -2px rgba(0,0,0,0.2),0px 2px 2px 0px rgba(0,0,0,0.14),0px 1px 5px 0px rgba(0,0,0,0.12)',
-  '0px 3px 3px -2px rgba(0,0,0,0.2),0px 3px 4px 0px rgba(0,0,0,0.14),0px 1px 8px 0px rgba(0,0,0,0.12)',
-  '0px 2px 4px -1px rgba(0,0,0,0.2),0px 4px 5px 0px rgba(0,0,0,0.14),0px 1px 10px 0px rgba(0,0,0,0.12)',
-  '0px 3px 5px -1px rgba(0,0,0,0.2),0px 5px 8px 0px rgba(0,0,0,0.14),0px 1px 14px 0px rgba(0,0,0,0.12)',
-  '0px 3px 5px -1px rgba(0,0,0,0.2),0px 6px 10px 0px rgba(0,0,0,0.14),0px 1px 18px 0px rgba(0,0,0,0.12)',
-  '0px 4px 5px -2px rgba(0,0,0,0.2),0px 7px 10px 1px rgba(0,0,0,0.14),0px 2px 16px 1px rgba(0,0,0,0.12)',
-  '0px 5px 5px -3px rgba(0,0,0,0.2),0px 8px 10px 1px rgba(0,0,0,0.14),0px 3px 14px 2px rgba(0,0,0,0.12)',
-];
-
-export const theme = {
-  type: 'light',
-  palette: palette,
-  spacing: spacing,
-  borderRadius: borderRadius,
-  typography: typography,
-  easing: easing,
-  shadows: shadows,
+export const getRemSize = (fontSize: number): number => {
+  const defaultFontSize: number = 14;
+  const htmlFontSize: number = 16;
+  const coefficient: number = defaultFontSize / 14;
+  return (fontSize / htmlFontSize) * coefficient;
 };
