+++ test/Button.test.tsx
@@ -4,7 +4,7 @@ import { Button } from '../src/';
 
 describe('Renders Button Component', () => {
   it('button is visible', () => {
-    render(<Button label={'Click Me'}>Click Me</Button>);
+    render(<Button $label={'Click Me'}>Click Me</Button>);
     expect(screen.getByRole('button')).toBeVisible();
     expect(screen.getByRole('button')).toBeInTheDocument();
   });
